# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Personal Finance API",
    "version": "1.0.0",
    "description": "Generated from Protobuf descriptors and registered HTTP routes. Standard Protobuf JSON. Private-network single-profile deployment."
  },
  "paths": {
    "/api/v1/categories": {
      "get": {
        "operationId": "get__api_v1_categories",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FinanceResponse" }
              }
            }
          },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post__api_v1_categories",
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Category" }
            }
          }
        }
      }
    },
    "/api/v1/categories/{id}": {
      "get": {
        "operationId": "get__api_v1_categories__id_",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ]
      },
      "patch": {
        "operationId": "patch__api_v1_categories__id_",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Category" }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete__api_v1_categories__id_",
        "responses": {
          "204": { "description": "Successful operation" },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": { "type": "string" },
            "description": "Current optimistic concurrency version."
          }
        ]
      }
    },
    "/api/v1/budgets": {
      "get": {
        "operationId": "get__api_v1_budgets",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FinanceResponse" }
              }
            }
          },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post__api_v1_budgets",
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Budget" }
              }
            }
          },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Budget" }
            }
          }
        }
      }
    },
    "/api/v1/budgets/{id}": {
      "get": {
        "operationId": "get__api_v1_budgets__id_",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Budget" }
              }
            }
          },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ]
      },
      "patch": {
        "operationId": "patch__api_v1_budgets__id_",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Budget" }
              }
            }
          },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Budget" }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete__api_v1_budgets__id_",
        "responses": {
          "204": { "description": "Successful operation" },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": { "type": "string" },
            "description": "Current optimistic concurrency version."
          }
        ]
      }
    },
    "/api/v1/recurring-commitments": {
      "get": {
        "operationId": "get__api_v1_recurring_commitments",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FinanceResponse" }
              }
            }
          },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post__api_v1_recurring_commitments",
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RecurringCommitment" }
              }
            }
          },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RecurringCommitment" }
            }
          }
        }
      }
    },
    "/api/v1/recurring-commitments/{id}": {
      "get": {
        "operationId": "get__api_v1_recurring_commitments__id_",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RecurringCommitment" }
              }
            }
          },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ]
      },
      "patch": {
        "operationId": "patch__api_v1_recurring_commitments__id_",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RecurringCommitment" }
              }
            }
          },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RecurringCommitment" }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete__api_v1_recurring_commitments__id_",
        "responses": {
          "204": { "description": "Successful operation" },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": { "type": "string" },
            "description": "Current optimistic concurrency version."
          }
        ]
      }
    },
    "/api/v1/categorization-rules": {
      "get": {
        "operationId": "get__api_v1_categorization_rules",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FinanceResponse" }
              }
            }
          },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post__api_v1_categorization_rules",
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CategorizationRule" }
              }
            }
          },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CategorizationRule" }
            }
          }
        }
      }
    },
    "/api/v1/categorization-rules/{id}": {
      "get": {
        "operationId": "get__api_v1_categorization_rules__id_",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CategorizationRule" }
              }
            }
          },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ]
      },
      "patch": {
        "operationId": "patch__api_v1_categorization_rules__id_",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CategorizationRule" }
              }
            }
          },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CategorizationRule" }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete__api_v1_categorization_rules__id_",
        "responses": {
          "204": { "description": "Successful operation" },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": { "type": "string" },
            "description": "Current optimistic concurrency version."
          }
        ]
      }
    },
    "/api/v1/scenarios": {
      "get": {
        "operationId": "get__api_v1_scenarios",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FinanceResponse" }
              }
            }
          },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post__api_v1_scenarios",
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Scenario" }
              }
            }
          },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Scenario" }
            }
          }
        }
      }
    },
    "/api/v1/scenarios/{id}": {
      "get": {
        "operationId": "get__api_v1_scenarios__id_",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Scenario" }
              }
            }
          },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ]
      },
      "patch": {
        "operationId": "patch__api_v1_scenarios__id_",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Scenario" }
              }
            }
          },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Scenario" }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete__api_v1_scenarios__id_",
        "responses": {
          "204": { "description": "Successful operation" },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": { "type": "string" },
            "description": "Current optimistic concurrency version."
          }
        ]
      }
    },
...
    "/api/v1/budgets/copy": {
      "post": {
        "operationId": "post__api_v1_budgets_copy",
        "responses": {
          "200": { "description": "Successful operation" },
          "default": {
            "description": "Structured application error",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/Problem" }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BudgetCopyRequest" }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FinanceResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "transactions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Transaction" }
          },
          "categories": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Category" }
          },
          "budgets": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Budget" }
          },
          "commitments": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/RecurringCommitment" }
          },
          "rules": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CategorizationRule" }
          },
          "exchangeRates": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ExchangeRate" },
            "deprecated": true,
            "description": "Legacy field retained for wire compatibility; ignored by the application."
          },
          "settings": { "$ref": "#/components/schemas/Settings" },
          "pagination": { "$ref": "#/components/schemas/Pagination" },
          "scenarios": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Scenario" }
          }
        }
      },
...
    }
  }
}